+++ client/app/scripts/liveblog-advertising/index.js
@@ -1,2 +1,5 @@
 import './styles/advertising.scss';
-import './module';
+import activities from './activities';
+
+var liveblogAdvertisingModule = angular.module('liveblog.advertising', [])
+.config(activities);
