+++ server/settings.py
@@ -138,7 +138,6 @@ INSTALLED_APPS = [
     'liveblog.blogslist',
     'liveblog.syndication',
     'liveblog.freetypes',
-    'liveblog.advertising',
     'liveblog.marketplace',
     'liveblog.analytics'
 ]
